Refactored ActiveModel::Type tests into their own files
