+++ activemodel/test/cases/type/binary_test.rb
@@ -1,0 +1,15 @@
+require "cases/helper"
+require "active_model/type"
+
+module ActiveModel
+  module Type
+    class BinaryTest < ActiveModel::TestCase
+      def test_type_cast_binary
+        type = Type::Binary.new
+        assert_equal nil, type.cast(nil)
+        assert_equal "1", type.cast("1")
+        assert_equal 1, type.cast(1)
+      end
+    end
+  end
+end
